pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "midi_xai"
version = "0.0.1"
description = "A short description of the project."
authors = [
  { name = "Zespol 4" },
]

readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    
]
dependencies = [
    "captum>=0.8.0",
    "hydra-core>=1.3.2",
    "lark>=1.3.1",
    "matplotlib>=3.10.8",
    "mkdocs",
    "pandas>=3.0.1",
    "partitura>=1.8.0",
    "pip",
    "pytest>=9.0.2",
    "python-dotenv",
    "ruff",
    "scikit-learn>=1.8.0",
    "torch>=2.10.0",
    "wandb>=0.25.1",
]
requires-python = "~=3.11.0"


[tool.ruff]
line-length = 99
src = ["midi_xai"]
include = ["pyproject.toml", "midi_xai/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["midi_xai"]
force-sort-within-sections = true

[dependency-groups]
dev = [
    "black>=26.3.1",
]

